UPDATE: network/adapter

+++ app/src/main/java/com/example/android/bakingapp/model/Ingredient.java
@@ -7,7 +7,7 @@ public class Ingredient {
 
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private Double quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -15,17 +15,17 @@ public class Ingredient {
     @Expose
     private String ingredient;
 
-    public Ingredient(Integer quantity, String measure, String ingredient) {
+    public Ingredient(Double quantity, String measure, String ingredient) {
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
     }
 
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
